Refactor benchmark utils and datasets
