+++ python/sglang/benchmark/__init__.py
@@ -1,0 +1,27 @@
+from . import datasets
+from .utils import (
+    download_and_cache_file,
+    download_and_cache_hf_file,
+    get_model,
+    get_processor,
+    get_tokenizer,
+    is_file_valid_json,
+    parse_custom_headers,
+    remove_prefix,
+    remove_suffix,
+    set_ulimit,
+)
+
+__all__ = [
+    "datasets",
+    "download_and_cache_file",
+    "download_and_cache_hf_file",
+    "get_model",
+    "get_processor",
+    "get_tokenizer",
+    "is_file_valid_json",
+    "parse_custom_headers",
+    "remove_prefix",
+    "remove_suffix",
+    "set_ulimit",
+]
